Add BotDialogue
