+++ examples/dialogue_bot/src/states.rs
@@ -1,7 +1,5 @@
 use teloxide::prelude::*;
 
-use super::transitions::*;
-
 #[derive(Default)]
 pub struct StartState;
 
@@ -30,20 +28,3 @@ up!(
     Receive10x5AnswerState + [_10x5_answer: u8] -> ReceiveGandalfAlternativeNameState,
     ReceiveGandalfAlternativeNameState + [gandalf_alternative_name: String] -> ExitState,
 );
-
-#[bot_dialogue]
-#[derive(SmartDefault, From)]
-pub enum Dialogue {
-    #[default]
-    #[handler(start)]
-    Start(StartState),
-
-    #[handler(receive_days_of_week)]
-    ReceiveDaysOfWeek(ReceiveDaysOfWeekState),
-
-    #[handler(receive_10x5_answer)]
-    Receive10x5Answer(Receive10x5AnswerState),
-
-    #[handler(receive_gandalf_alternative_name)]
-    ReceiveGandalfAlternativeName(ReceiveGandalfAlternativeNameState),
-}
