+++ examples/dialogue_bot/src/transitions.rs
@@ -55,3 +55,19 @@ async fn receive_gandalf_alternative_name(
         }
     }
 }
+
+#[derive(BotDialogue, SmartDefault, From)]
+pub enum Dialogue {
+    #[default]
+    #[handler(start)]
+    Start(StartState),
+
+    #[handler(receive_days_of_week)]
+    ReceiveDaysOfWeek(ReceiveDaysOfWeekState),
+
+    #[handler(receive_10x5_answer)]
+    Receive10x5Answer(Receive10x5AnswerState),
+
+    #[handler(receive_gandalf_alternative_name)]
+    ReceiveGandalfAlternativeName(ReceiveGandalfAlternativeNameState),
+}
